Provide mock meter server and validate (#25)
* Add meter mock collector and validate

* Add comments and NPE check

* Add data description

+++ validator/src/main/java/org/apache/skywalking/plugin/test/agent/tool/validator/assertor/element/GreatThanAssertor.java
@@ -27,7 +27,7 @@ public class GreatThanAssertor extends ElementAssertor {
 
     @Override
     public void assertValue(String desc, String actualValue) {
-        if (Long.parseLong(actualValue) <= Long.parseLong(exceptedValue)) {
+        if (Double.parseDouble(actualValue) <= Double.parseDouble(exceptedValue)) {
             throw new ValueAssertFailedException(desc, " gt " + exceptedValue, actualValue);
         }
     }
